requirements.txt
aiohttp==3.6.3
astroid==2.5.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.4
discord-components==2.0.4
discord-ext-ui==2.1.2
discord-ui==4.3.8
discord.py==1.5.1
googletrans==3.0.0
gTTS==2.2.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.20
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==4.7.6
openai==0.27.2
pycparser==2.20
pydub==0.25.1
pylint==2.7.4
PyNaCl==1.3.0
pytz==2021.1
requests==2.28.2
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
toml==0.10.2
tornado==4.5.3
tqdm==4.65.0
urllib3==1.25.11
wincertstore==0.2
wrapt==1.12.1
yarl==1.5.1
